code refactor: move interfaces to types.ts
